#629 system tests
