+++ tb-gcp-tools/tb-auto-projects-deleter/cloud-function/config.py
@@ -46,8 +46,9 @@ def str_to_list(string_list: str):
 dry_run = str_to_bool(get_env('dry_run', True))
 is_cloud_run = str_to_bool(get_env('is_cloud_run', "False"))
 EXCLUDE_DELETE_LABEL = get_env('EXCLUDE_DELETE_LABEL', "dont-delete")
-EXCLUDE_EMPTY_FOLDER = str_to_list(get_env('EXCLUDE_EMPTY_FOLDER', "Shared Services,  Applications"))
-ROOT_PROJECT = get_env('ROOT_PROJECT', "943956663445")
+EXCLUDE_EMPTY_FOLDER = str_to_list(get_env('EXCLUDE_EMPTY_FOLDER', "Shared Services,  Applications, tranquilitybase-development"))
+ROOT_PROJECT = get_env('ROOT_PROJECT', "256793750708")
+TEST_OUTSIDE_SCOPE = get_env('TEST_OUTSIDE_SCOPE', "261170752178")
 ReportWebhook = get_env('WebHook', "none")
 
 
@@ -57,6 +58,7 @@ print("dry_run: " + str(dry_run))
 print("is_cloud_run: " + str(is_cloud_run))
 print("EXCLUDE_DELETE_LABEL: " + EXCLUDE_DELETE_LABEL)
 print("EXCLUDE_EMPTY_FOLDER: " + str(EXCLUDE_EMPTY_FOLDER))
+print("TEST_OUTSIDE_SCOPE: " + str(TEST_OUTSIDE_SCOPE))
 print("ROOT_PROJECT: " + ROOT_PROJECT)
 print("---------------------------")
 print("")
