+++ tb-gcp-tools/tb-auto-projects-deleter/cloud-function/services/helper.py
@@ -1,0 +1,6 @@
+def handle_permission_error(exception: Exception, message: str):
+    if "not have permission" in str(exception):
+        print(str(exception))
+        print(message)
+    else:
+        raise exception
